feat: add payment and update thank pages
